pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-paper-review"
version = "0.3.1"
description = "AI paper review system: multiple LLM reviewer personas, parallel review, clustering and ranking, human-feedback calibration."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Paper Review Contributors" }]
keywords = ["peer-review", "llm", "computer-architecture", "langgraph"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Framework :: Flask",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "langgraph>=0.2.0",
    "pypdf>=4.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "sentence-transformers>=2.7",
    "flask>=3.0",
    "markdown>=3.4",
    "pyyaml>=6.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-genai>=0.3.0",
    "github-copilot-sdk>=0.1.0",
    "claude-agent-sdk>=0.1.0",
    "markitdown[pdf]>=0.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pyflakes>=3.0",
    "build>=1.0",
]

[project.scripts]
ai-paper-review-review    = "ai_paper_review.review.review:main"
ai-paper-review-validate  = "ai_paper_review.validation.validation:main"
ai-paper-review-aggregate = "ai_paper_review.aggregation.aggregation:main"
ai-paper-review-web       = "ai_paper_review.web.app:main"

[project.urls]
Homepage   = "https://github.com/diwu1990/ai-paper-review"
Repository = "https://github.com/diwu1990/ai-paper-review"
Issues     = "https://github.com/diwu1990/ai-paper-review/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_paper_review"]
# Hatch's package discovery already includes every file under
# `src/ai_paper_review/` — including the bundled `data/*.md`,
# `data/*.yaml`, and `prompts/*.md`. Do NOT add a
# `[tool.hatch.build.targets.wheel.force-include]` block re-targeting
# those paths: PyPI now rejects wheels with duplicate ZIP entries
# (https://docs.pypi.org/project_metadata/#duplicate-filenames-in-zip-archives).

[tool.hatch.build.targets.sdist]
include = [
    "src/ai_paper_review",
    "examples",
    "docs",
    "tests",
    "config.example.yaml",
    "environment.yml",
    "README.md",
    "LICENSE",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
